empty address layout added

+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/ui/fragments/AddressFragment.kt
@@ -16,9 +16,6 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.getflix.R
 import com.example.getflix.databinding.AddressCardItemBinding
 import com.example.getflix.databinding.FragmentAddressBinding
-import com.example.getflix.databinding.FragmentProfileBinding
-import com.example.getflix.models.AddressModel
-import com.example.getflix.models.PhoneModel
 import com.example.getflix.ui.adapters.*
 import com.example.getflix.ui.fragments.AddressFragmentDirections.Companion.actionAddressFragmentToAddAddressFragment
 import com.example.getflix.ui.viewmodels.AddressViewModel
@@ -41,7 +38,13 @@ class AddressFragment : Fragment() {
         binding = DataBindingUtil.inflate(inflater, R.layout.fragment_address,
             container, false)
         binding.fab.setOnClickListener {
+            if(binding.fab.visibility == View.VISIBLE)
             view?.findNavController()?.navigate(actionAddressFragmentToAddAddressFragment())
+        }
+
+        binding.btnAddAddress.setOnClickListener {
+            if(binding.btnAddAddress.visibility == View.VISIBLE)
+                view?.findNavController()?.navigate(actionAddressFragmentToAddAddressFragment())
         }
 
         viewModel = ViewModelProvider(this).get(AddressViewModel::class.java)
@@ -53,20 +56,34 @@ class AddressFragment : Fragment() {
 
         viewModel.addressList.observe(viewLifecycleOwner, Observer {
             it?.let {
-                val addressListAdapter = AddressAdapter(ArrayList(it!!),this)
-                recView.adapter = addressListAdapter
-                recView.setHasFixedSize(true)
-               // addressListAdapter.submitList(it)
-                        val itemTouchHelper = ItemTouchHelper(SwipeToDeleteAddress(addressListAdapter!!))
-                        itemTouchHelper.attachToRecyclerView(recView)
-                        addressListAdapter.pos.observe(viewLifecycleOwner, Observer {
-                            if (it != -1) {
-                                val id = addressListAdapter.deleteItem(it).id
-                                viewModel.deleteCustomerAddress(id)
-                                addressListAdapter.resetPos()
-                            }
-                        })
-                    }
+                if (it.size != 0) {
+                    binding.btnAddAddress.visibility = View.GONE
+                    binding.addressImage.visibility = View.GONE
+                    binding.addressText.visibility = View.GONE
+                    binding.addressList.visibility = View.VISIBLE
+                    binding.fab.visibility = View.VISIBLE
+                    val addressListAdapter = AddressAdapter(ArrayList(it!!), this)
+                    recView.adapter = addressListAdapter
+                    recView.setHasFixedSize(true)
+                    // addressListAdapter.submitList(it)
+                    val itemTouchHelper =
+                        ItemTouchHelper(SwipeToDeleteAddress(addressListAdapter!!))
+                    itemTouchHelper.attachToRecyclerView(recView)
+                    addressListAdapter.pos.observe(viewLifecycleOwner, Observer {
+                        if (it != -1) {
+                            val id = addressListAdapter.deleteItem(it).id
+                            viewModel.deleteCustomerAddress(id)
+                            addressListAdapter.resetPos()
+                        }
+                    })
+                } else {
+                    binding.btnAddAddress.visibility = View.VISIBLE
+                    binding.addressImage.visibility = View.VISIBLE
+                    binding.addressText.visibility = View.VISIBLE
+                    binding.addressList.visibility = View.GONE
+                    binding.fab.visibility = View.GONE
+                }
+            }
                 })
